fix(model): altera parâmetro de setAge()

+++ src/main/java/com/devbigode/petadoption/model/Pet.java
@@ -38,7 +38,7 @@ public class Pet {
         return age;
     }
 
-    public void setAge(int age) {
+    public void setAge(double age) {
         this.age = age;
     }
